Luca's skin provider.

+++ contact-modelling/src/AlgDisplacementsToNonnegativePressures.cpp
@@ -108,7 +108,7 @@ void AlgDisplacementsToNonnegativePressures::impl_run(
   LOG(DEBUG) << "nnls residual norm: " << residualNorm;
   pressures.getRawValues().clear();
   pressures.getRawValues().reserve(pre.taucs_m->n);
-  for (size_t i = 0; i < pre.taucs_m->n; ++i) {
+  for (size_t i = 0; i < (size_t)pre.taucs_m->n; ++i) {
     pressures.getRawValues().push_back(*(solution+i));
   }
   free(solution);
